+++ CHANGELOG.md
@@ -12,11 +12,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 ### Deprecated
+- Deprecated the following methods: \OxidEsales\PayPalModule\Core\Config::getIPNResponseUrl()
 
 ### Removed
 
 ### Fixed
 - Compatibility of tests with MySQL 5.7.
+- Fixed 0006122 IPN postback DNS issue. Introduced \OxidEsales\PayPalModule\Core\IpnConfig class to 
+  provide the necessary IPN parameters for host and url.  
 
 ### Security